Get rid of the `Stake` page
Rename the `Stake` page to `Mint`. On the page itself, remove everything
below the `Why mint tBTC?` section. Also create a redirect in gatsby
config file that redirects from `/stake` to `/mint` page just in case
someone has the `/stake` page in bookmark.
